+++ main.go
@@ -1,0 +1,5 @@
+// Sample main.go
+
+package main
+
+func main() {}